requirements.txt
python-telegram-bot
requests
google-generativeai